feat: add `backtrace` syscall
The original lab called for adding the backtrace call in `sys_sleep`,
but I opted to make it a dedicated syscall, since it felt wrong to
"break" sleep.

Since I was adding a new syscall and syscall numbers up to 30 (for
future labs) were already defined, I set this up at 31.