feat: implement copy-on-write `fork()`
This commit optimizes `fork()` by changing it to no longer copy all of a
process's pages when creating the child. Instead, `uvmcopy` shares the
pages between pages and child, and uses the newly implemented `PTE_COW`
bit for writable pages to indicate that a write-fault should trigger an
on-demand copy.

Since the parent and child processes now share pages, we must also keep
track of the number of reference to a given page, so that when we clean
up pages for a given process, we don't free pages that are being used by
other processes. I addressed this with a contiguous array of counts,
with one entry for each physical page, in `kalloc.c`. If I understand
correctly, the Linux kernel has a contiguous array of _structs_ to store
page metadata, so this is something we could explore later on. [^1]

I removed `freerange` and added a new `kuncheckedfree` function to
simplify page refcount handling during initialization. A nice
side-effect of this refactoring is we don't need to memset() all of
physical memory during boot, since `kuncheckedfree` doesn't wipe the
memory it frees.

The last two changes to `kalloc.c` are also refcount-related: the first,
`kincrementrefcount`, allows `uvmcopy` to indicate that multiple
processes are sharing a single page. The second, `kcopyonwrite`,
provides a simple interface to allocate a new page on demand for a given
copy-on-write page. It is optimized to *not* perform an allocation when
the caller is the last remaining reference to a page, instead opting to
return the page back to the caller.

There are two places in the kernel where we need to perform an on-demand
copy of a copy-on-write page: in `usertrap`, when a userspace process
attempts to write to a page marked as PTE_COW, and in `copyout`, when
the kernel attempts to write to a userspace address in a page marked as
PTE_COW.

I implemented `uvmwalkcow` to centralize the logic for this. It uses the
aforementioned `kcopyonwrite` function for PTE_COW pages to
transparently allocate a new physical page for a given virtual address,
and toggles the PTE_COW and PTE_W bits so the caller has a writable
page. `usertrap` and `copyout` simply ask `uvmwalkcow` for a pte_t, and
can react appropriately with the result.

Finally, I decided to refactor `usertrap` during the process of
implementing the necessary changes for copy-on-write. I prefer functions
with shallow conditionals that return early, and so I took a stab at
organizing `usertrap` to be somewhat more Golang-like, and I think the
result is, while not perfect, more readable than before. I also
introduced new constants, like SCAUSE_ECALL_UMODE, to avoid "magic"
numbers.

[^1]: https://github.com/lorenzo-stoakes/linux-vm-notes/blob/master/sections/page-tables.md