feat: add `mmap` and `munmap`
This commit adds a somewhat, but not entirely, primitive implementation
of `mmap` and `munmap`.

The lab suggested using a statically-allocated array of vm_area structs
instead of dynamically allocating memory for them, so I went ahead and
made an array of size NPROC.

The lab didn't require supporting user-specified addresses, arbitrary
file offsets, nor unmapping regions in the middle of a VMA, but I ended
up supporting the latter two as it was relatively straightforward when
following the algorithm in Linux's `do_vmi_align_munmap`[^1].

This ended up being very helpful when implementing `munmap`, even though
Linux uses a data structure more advanced than a linked list. The core
algorithm seems to roughly be:

1. Find the first VMA that overlaps with the unmapped address.
2. Split this VMA if the unmapped address is larger than the VMA's
   starting address.
3. Iterate over each VMA, starting at the VMA from 1 until the last VMA
   that contains the end address to unmap.
4. If the current VMA extends past the end endress, split this VMA. In
   practice this should only ever apply to the last VMA.
5. Free the current VMA.

Splitting a VMA is straightforward, but requires modifying the file
offset of the split VMA, which is why my implementation supports
arbitrary file offsets in `mmap`.

I had originally tried to follow the lab's suggestion to not split VMAs,
but I found it difficult to keep track of when a VMA needed to be fully
deallocated, and when its backing file and pages should be closed and
unmapped. Splitting the VMAs made this simpler, as I would always
completely free any VMAs that were contained in the unmapped range, and
leave the split ones intact.

[^1]: https://elixir.bootlin.com/linux/v6.8.1/source/mm/mmap.c#L2512